pyproject.toml
[tool.poetry]
name = "habit-track-bot"
version = "0.1.0"
description = ""
authors = ["Vladislav Grabinski <bull2099@yandex.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
alembic = "1.18.4"
annotated-doc = "0.0.4"
annotated-types = "0.7.0"
anyio = "4.12.1"
apscheduler = "3.11.2"
asyncpg = "0.31.0"
bcrypt = "5.0.0"
certifi = "2026.1.4"
charset-normalizer = "3.4.4"
click = "8.3.1"
docker = "7.1.0"
fastapi = "0.129.0"
greenlet = "3.3.1"
h11 = "0.16.0"
idna = "3.11"
mako = "1.3.10"
markupsafe = "3.0.3"
passlib = "1.7.4"
psycopg2-binary = "2.9.11"
psycopg2-pool = "1.2"
pydantic = "2.12.5"
pydantic-core = "2.41.5"
pyjwt = "2.11.0"
pytelegrambotapi = "4.31.0"
python-dotenv = "1.2.1"
requests = "2.32.5"
sqlalchemy = "2.0.46"
starlette = "0.52.1"
typing-inspection = "0.4.2"
typing-extensions = "4.15.0"
tzdata = "2025.3"
tzlocal = "5.3.1"
urllib3 = "2.6.3"
uvicorn = "0.41.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.alembic]
script_location = "db/migrations"
prepend_sys_path = ".."
sqlalchemy.url = "postgresql+asyncpg://postgres:mysecretpassword@postgres:5432/habit_bot_tg"
